pyproject.toml
[tool.pytest.ini_options]
testpaths = [
    "portal",
    "tests"
]

[project]
name = "conf-portal-api"
version = "0.1.0"
description = ""
authors = [{ name = "Jay Hsia", email = "jayhsia1997@gmail.com" }]
readme = "README.md"
package-mode = false
requires-python = "^3.13"
dependencies = [
    # base
    "dependency-injector",
    "fastapi",
    "fastapi-limiter==0.1.6",
    "gunicorn",
    "uvicorn (==0.35.0)",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    # db
    "asyncpg",
    "redis",
    "sqlalchemy",
    "psycopg2-binary==2.9.10",
    # firebase
    "firebase-admin (==7.1.0)",
    # security
    "pyjwt",
    # utils
    "boto3",
    "cryptography",
    "email-validator",
    "httpx",
    "ujson",
    "pytz",
    "pandas (==2.3.3)",
    "numpy (==2.3.4)",
    "phonenumbers",
    "pyotp",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.20",
    "sentry-sdk[fastapi]",
    "pillow (>=11.3.0,<12.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
alembic = "*"
pytest = "*"
pytest-asyncio = "*"
pytest-mock = "*"
click = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
